Fixed Ubuntu version check
This code works with Python 2 and 3. Only the Python 3 version of os.uname()
returns an object with the version attribute - in Python 2 it's just a tuple.
But for backwards compatibility, the Python 3 object is also iterable, behaving
like a five-tuple containing sysname, nodename, release, version, and machine
in that order.